requirements.txt
alembic==1.16.5
altair==4.2.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==26.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
cloudpickle==3.1.2
colorlog==6.10.1
contourpy==1.3.0
cryptography==46.0.5
cycler==0.12.1
databricks-sdk==0.102.0
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
dynaconf==3.2.13
entrypoints==0.4
evidently==0.4.30
exceptiongroup==1.3.1
Faker==37.12.0
fastapi==0.128.8
filelock==3.19.1
Flask==3.1.3
fonttools==4.60.2
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
great_expectations==1.8.1
groq==0.37.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
langchain==0.3.28
langchain-core==0.3.83
langchain-groq==0.3.8
langchain-text-splitters==0.3.11
langsmith==0.4.37
lightgbm==4.6.0
litestar==2.21.1
litestar-htmx==0.5.0
llvmlite==0.43.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.9.4
mdurl==0.1.2
mistune==3.2.0
mlflow==3.1.4
mlflow-skinny==3.1.4
msgspec==0.20.0
multidict==6.7.1
multipart==1.3.1
mypy_extensions==1.1.0
nltk==3.9.2
numba==0.60.0
numpy==2.0.2
opentelemetry-api==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
optuna==4.8.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pillow==11.3.0
plotly==5.24.1
polyfactory==3.3.0
protobuf==6.33.6
pyarrow==20.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rich-click==1.9.7
rpds-py==0.27.1
ruamel.yaml==0.19.1
scikit-learn==1.6.1
scipy==1.13.1
shap==0.49.1
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlparse==0.5.5
starlette==0.49.3
statsmodels==0.14.6
streamlit==1.50.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.4.0
toolz==1.1.0
tornado==6.5.5
tqdm==4.67.3
typer==0.23.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.3
uuid6==2025.0.1
uuid_utils==0.14.1
uvicorn==0.39.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.6
xgboost==2.1.4
zipp==3.23.0
zstandard==0.25.0
